Replace userSuggestions with react-query hook / remove duplications
